[Feature WIP] Update Book price's type to float

+++ src/Entity/Book.php
@@ -107,12 +107,12 @@ class Book
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
